refactor: rename lib/collective.lib -> lib/collective
