+++ components/AccountHoverCard.tsx
@@ -6,7 +6,7 @@ import { get } from 'lodash';
 import { Banknote, Building, Calendar, FileText, LucideIcon, Mail, PencilRuler, Receipt, Users } from 'lucide-react';
 import { FormattedDate, FormattedMessage } from 'react-intl';
 
-import { isIndividualAccount } from '../lib/collective.lib';
+import { isIndividualAccount } from '../lib/collective';
 import { API_V2_CONTEXT, gql } from '../lib/graphql/helpers';
 import { Account, AccountWithHost, UserContextualMembershipsQuery } from '../lib/graphql/types/v2/graphql';
 import { getCollectivePageRoute } from '../lib/url-helpers';
