+++ components/CreateOrganizationForm.js
@@ -6,7 +6,7 @@ import { withRouter } from 'next/router';
 import { defineMessages, FormattedMessage, injectIntl } from 'react-intl';
 import { isURL } from 'validator';
 
-import { suggestSlug } from '../lib/collective.lib';
+import { suggestSlug } from '../lib/collective';
 
 import { BackButton } from './create-collective/CreateCollectiveForm';
 import OnboardingProfileCard from './onboarding-modal/OnboardingProfileCard';
