+++ components/admin-panel/sections/ActivityLog/ActivityTypeFilter.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { omitBy } from 'lodash';
 import { defineMessage, useIntl } from 'react-intl';
 
-import { isIndividualAccount } from '../../../../lib/collective.lib';
+import { isIndividualAccount } from '../../../../lib/collective';
 import { ActivityTypes } from '../../../../lib/constants/activities';
 import { ActivityTypeI18n } from '../../../../lib/i18n/activities';
