+++ components/collective-navbar/ActionsMenu.js
@@ -13,7 +13,7 @@ import { pickBy } from 'lodash';
 import { FormattedMessage } from 'react-intl';
 import styled, { css } from 'styled-components';
 
-import { getContributeRoute } from '../../lib/collective.lib';
+import { getContributeRoute } from '../../lib/collective';
 import { getCollectivePageRoute, getSettingsRoute } from '../../lib/url-helpers';
 
 import ActionButton from '../ActionButton';
