+++ components/collective-page/sections/Location.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
-import { isEmptyCollectiveLocation } from '../../../lib/collective.lib';
+import { isEmptyCollectiveLocation } from '../../../lib/collective';
 import useLoggedInUser from '../../../lib/hooks/useLoggedInUser';
 
 import Container from '../../Container';
