+++ components/expenses/ExpenseFormPayeeInviteNewStep.js
@@ -6,7 +6,7 @@ import { get, isEmpty, omit } from 'lodash';
 import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
 import styled from 'styled-components';
 
-import { suggestSlug } from '../../lib/collective.lib';
+import { suggestSlug } from '../../lib/collective';
 import { EMPTY_ARRAY } from '../../lib/constants/utils';
 import { ERROR, isErrorType } from '../../lib/errors';
 import { formatFormErrorMessage, requireFields, verifyEmailPattern } from '../../lib/form-utils';
