+++ components/ocf-host-application/ApplicationForm.js
@@ -10,7 +10,7 @@ import { get, isNil } from 'lodash';
 import { withRouter } from 'next/router';
 import { FormattedMessage, useIntl } from 'react-intl';
 
-import { suggestSlug } from '../../lib/collective.lib';
+import { suggestSlug } from '../../lib/collective';
 import { OPENCOLLECTIVE_FOUNDATION_ID } from '../../lib/constants/collectives';
 import { formatCurrency } from '../../lib/currency-utils';
 import { i18nGraphqlException } from '../../lib/errors';
