+++ lib/__tests__/collective.js
@@ -1,6 +1,6 @@
 import { CollectiveCategory, OPENSOURCE_COLLECTIVE_ID } from '../../lib/constants/collectives';
 
-import { getCollectiveMainTag } from '../collective.lib';
+import { getCollectiveMainTag } from '../collective';
 
 const UNKNOWN_HOST_ID = 42;
