+++ lib/collective-sections.js
@@ -17,7 +17,7 @@ import { Sections } from '../components/collective-page/_constants';
 import { CollectiveType } from './constants/collectives';
 import i18nNavbarCategory from './i18n/navbar-categories';
 import { FEATURES, isFeatureEnabled } from './allowed-features';
-import { isEmptyCollectiveLocation } from './collective.lib';
+import { isEmptyCollectiveLocation } from './collective';
 
 import aboutCategoryIcon from '../public/static/images/collective-navigation/CollectiveNavbarIconAbout.png';
 import budgetCategoryIcon from '../public/static/images/collective-navigation/CollectiveNavbarIconBudget.png';
