+++ lib/collective.js
