+++ pages/collective-contact.tsx
@@ -3,7 +3,7 @@ import { useQuery } from '@apollo/client';
 import { useRouter } from 'next/router';
 import { FormattedMessage, IntlShape, useIntl } from 'react-intl';
 
-import { getCollectivePageMetadata } from '../lib/collective.lib';
+import { getCollectivePageMetadata } from '../lib/collective';
 import { generateNotFoundError } from '../lib/errors';
 import { API_V2_CONTEXT, gql } from '../lib/graphql/helpers';
 import { Account } from '../lib/graphql/types/v2/graphql';
