+++ pages/order.tsx
@@ -6,7 +6,7 @@ import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
 import styled from 'styled-components';
 
 import { getSSRQueryHelpers } from '../lib/apollo-client';
-import { getCollectivePageMetadata } from '../lib/collective.lib';
+import { getCollectivePageMetadata } from '../lib/collective';
 import dayjs from '../lib/dayjs';
 import { API_V2_CONTEXT, gql } from '../lib/graphql/helpers';
 import { Account, AccountWithHost } from '../lib/graphql/types/v2/graphql';
